Create the clearCompletedTasks function in checkbox module and call from index.js
