+++ src/module/constElements.js
@@ -2,17 +2,9 @@ const refreshTask = document.querySelector('#refresh');
 const taskInput = document.querySelector('#task');
 const submitInput = document.querySelector('#arrow-submit');
 const taskList = document.querySelector('#taskList');
-const clearCompletedTasks = document.querySelector('#clear-completed');
+// const clearCompletedTasks = document.querySelector('#clear-completed');
 
 const threeDots = document.querySelector('.three-dots');
 const trash = document.querySelector('.trash');
 
-export {
-  refreshTask,
-  taskInput,
-  submitInput,
-  taskList,
-  clearCompletedTasks,
-  threeDots,
-  trash,
-};
+export { refreshTask, taskInput, submitInput, taskList, threeDots, trash };
