+++ src/styles/style.css
@@ -47,7 +47,7 @@ span:focus {
 /* input[type='checkbox']:checked + label.strikethrough {
   text-decoration: line-through;
 } */
-
+/* 
 input[type='checkbox']:checked + label.strikethrough {
   text-decoration: line-through;
-}
+} */
